FS: fix defect related to grammar
The defect was related to identifying location when parseTree has only
functionChain or value.
Now, the location is clearly identified for expressionGroup,
functionChain and also value types.

Only with above defect, another defect: saving filter fails when the
expression is repeated. Because, it was trying to save duplicate entity
in ExpressionEntity. This is fixed, by ensuring only unique entity is
saved.